requirements.txt
onnxruntime==1.18.0
category-encoders==2.6.3
scikit-learn==1.2.2
numpy==1.26.4
pandas==2.1.4
streamlit==1.30.0
pillow==10.2.0